split handler impl code
