+++ pkg/agent/core/ngt/handler/grpc/index.go
@@ -1,0 +1,153 @@
+package grpc
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/vdaas/vald/apis/grpc/v1/payload"
+	"github.com/vdaas/vald/internal/errors"
+	"github.com/vdaas/vald/internal/info"
+	"github.com/vdaas/vald/internal/log"
+	"github.com/vdaas/vald/internal/net/grpc/errdetails"
+	"github.com/vdaas/vald/internal/net/grpc/status"
+	"github.com/vdaas/vald/internal/observability/trace"
+)
+
+func (s *server) CreateIndex(ctx context.Context, c *payload.Control_CreateIndexRequest) (res *payload.Empty, err error) {
+	ctx, span := trace.StartSpan(ctx, apiName+".CreateIndex")
+	defer func() {
+		if span != nil {
+			span.End()
+		}
+	}()
+	res = new(payload.Empty)
+	err = s.ngt.CreateIndex(ctx, c.GetPoolSize())
+	if err != nil {
+		if errors.Is(err, errors.ErrUncommittedIndexNotFound) {
+			err = status.WrapWithFailedPrecondition(fmt.Sprintf("CreateIndex API failed to create indexes pool_size = %d", c.GetPoolSize()), err,
+				&errdetails.RequestInfo{
+					ServingData: errdetails.Serialize(c),
+				},
+				&errdetails.ResourceInfo{
+					ResourceType: ngtResourceType + "/ngt.CreateIndex",
+					ResourceName: fmt.Sprintf("%s: %s(%s)", apiName, s.name, s.ip),
+				},
+				&errdetails.PreconditionFailure{
+					Violations: []*errdetails.PreconditionFailureViolation{
+						{
+							Type:    "uncommited index is empty",
+							Subject: "failed to CreateIndex operation caused by empty uncommited indices",
+						},
+					},
+				}, info.Get())
+			if span != nil {
+				span.SetStatus(trace.StatusCodeFailedPrecondition(err.Error()))
+			}
+			return nil, err
+		}
+		log.Error(err)
+		err = status.WrapWithInternal(fmt.Sprintf("CreateIndex API failed to create indexes pool_size = %d", c.GetPoolSize()), err,
+			&errdetails.RequestInfo{
+				ServingData: errdetails.Serialize(c),
+			},
+			&errdetails.ResourceInfo{
+				ResourceType: ngtResourceType + "/ngt.CreateIndex",
+				ResourceName: fmt.Sprintf("%s: %s(%s)", apiName, s.name, s.ip),
+			}, info.Get())
+		log.Error(err)
+		if span != nil {
+			span.SetStatus(trace.StatusCodeInternal(err.Error()))
+		}
+		return nil, err
+	}
+	return res, nil
+}
+
+func (s *server) SaveIndex(ctx context.Context, _ *payload.Empty) (res *payload.Empty, err error) {
+	ctx, span := trace.StartSpan(ctx, apiName+".SaveIndex")
+	defer func() {
+		if span != nil {
+			span.End()
+		}
+	}()
+	res = new(payload.Empty)
+	err = s.ngt.SaveIndex(ctx)
+	if err != nil {
+		log.Error(err)
+		err = status.WrapWithInternal("SaveIndex API failed to save indices", err,
+			&errdetails.ResourceInfo{
+				ResourceType: ngtResourceType + "/ngt.SaveIndex",
+				ResourceName: fmt.Sprintf("%s: %s(%s)", apiName, s.name, s.ip),
+			}, info.Get())
+		log.Error(err)
+		if span != nil {
+			span.SetStatus(trace.StatusCodeInternal(err.Error()))
+		}
+		return nil, err
+	}
+	return res, nil
+}
+
+func (s *server) CreateAndSaveIndex(ctx context.Context, c *payload.Control_CreateIndexRequest) (res *payload.Empty, err error) {
+	ctx, span := trace.StartSpan(ctx, apiName+".CreateAndSaveIndex")
+	defer func() {
+		if span != nil {
+			span.End()
+		}
+	}()
+	res = new(payload.Empty)
+	err = s.ngt.CreateAndSaveIndex(ctx, c.GetPoolSize())
+	if err != nil {
+		if errors.Is(err, errors.ErrUncommittedIndexNotFound) {
+			err = status.WrapWithFailedPrecondition(fmt.Sprintf("CreateAndSaveIndex API failed to create indexes pool_size = %d", c.GetPoolSize()), err,
+				&errdetails.RequestInfo{
+					ServingData: errdetails.Serialize(c),
+				},
+				&errdetails.ResourceInfo{
+					ResourceType: ngtResourceType + "/ngt.CreateAndSaveIndex",
+					ResourceName: fmt.Sprintf("%s: %s(%s)", apiName, s.name, s.ip),
+				},
+				&errdetails.PreconditionFailure{
+					Violations: []*errdetails.PreconditionFailureViolation{
+						{
+							Type:    "uncommited index is empty",
+							Subject: "failed to CreateAndSaveIndex operation caused by empty uncommited indices",
+						},
+					},
+				}, info.Get())
+			if span != nil {
+				span.SetStatus(trace.StatusCodeFailedPrecondition(err.Error()))
+			}
+			return nil, err
+		}
+		err = status.WrapWithInternal(fmt.Sprintf("CreateAndSaveIndex API failed to create indexes pool_size = %d", c.GetPoolSize()), err,
+			&errdetails.RequestInfo{
+				ServingData: errdetails.Serialize(c),
+			},
+			&errdetails.ResourceInfo{
+				ResourceType: ngtResourceType + "/ngt.CreateAndSaveIndex",
+				ResourceName: fmt.Sprintf("%s: %s(%s)", apiName, s.name, s.ip),
+			}, info.Get())
+		log.Error(err)
+		if span != nil {
+			span.SetStatus(trace.StatusCodeInternal(err.Error()))
+		}
+		return nil, err
+	}
+	return res, nil
+}
+
+func (s *server) IndexInfo(ctx context.Context, _ *payload.Empty) (res *payload.Info_Index_Count, err error) {
+	ctx, span := trace.StartSpan(ctx, apiName+".IndexInfo")
+	defer func() {
+		if span != nil {
+			span.End()
+		}
+	}()
+	return &payload.Info_Index_Count{
+		Stored:      uint32(s.ngt.Len()),
+		Uncommitted: uint32(s.ngt.InsertVQueueBufferLen() + s.ngt.DeleteVQueueBufferLen()),
+		Indexing:    s.ngt.IsIndexing(),
+		Saving:      s.ngt.IsSaving(),
+	}, nil
+}
